Refactored to make frontend dynamic
